Lock Powerup Update
Added a locked Brick to the level spawning, as well as a key powerup.
Locked Brick is not breakable by the ball normally, unless they have the key Powerup!
Key Powerup spawns randomly and descends toward the bottom of the screen, where the Paddle has the chance to collide with it and pick it up.